Fix Code Coverage Issue
Found that when running the Code Coverage tool EclEmma some tests were
unexpectedly failing. Tracked down to existence of unexpected field
(boolean[] jacocoData) present in the classes. This field is used by the
Coverage tool to store the coverage information for the class.  Added
legitimate workaround to rectify the problem.